Homework 6

+++ app/src/main/java/com/dbendyug/loftcoin/util/ImageURLFormatter.java
@@ -2,5 +2,5 @@ package com.dbendyug.loftcoin.util;
 
 public interface ImageURLFormatter {
 
-    String format(int id);
+    String format(long id);
 }
